requirements.txt
discord
python-dotenv
numpy
matplotlib
tensorflow
Pillow
opencage
pandas
meteomatics
